backend/requirements.txt
fastapi
uvicorn
numpy
tensorflow
pillow
google-generativeai
